Clean up the features.
