+++ features/README.md
@@ -1,8 +1,21 @@
 This is an experimental feature suite in Gherkin, to write Specifications by Example.
 It's a work in progress, and there's a lot of work ahead. Write a feature, and dive in !
 
-Ideally running the feature suite should not ruin the dev database, (unless we want it to)
-and it should be done using a meteor command such as test-features.
+``` gherkin
+Feature: Writing gherkin scenarios
+In order to describe the feature I want
+As a contributor
+I want to be able to write scenarios and steps
+
+Feature: Running gherkin scenarios
+In order to develop quickly and fearlessly
+As a developer
+I want to be able to run the feature suite
+```
+
+Ideally running the feature suite should not ruin the dev database, _(unless we want it to)_
+and it should be done using a meteor command such as `test-features`.
+
 
 ## WARNING
 
@@ -10,6 +23,7 @@ Running the feature suite will TRUNCATE THE DATABASE before each Scenario.
 You will lose everything you had on the development server.
 At the end of the test suite, the very last Scenario is not cleared,
 so we can use that to create a fixtures Scenario to bootstrap local development.
+
 
 ## Roadmap
 
@@ -23,13 +37,14 @@ so we can use that to create a fixtures Scenario to bootstrap local development.
 - [ ] Debating
     - [ ] Using a chronological tree
 
+
 ## Run
 
 Launch the development server :
 
     $ meteor npm run start:dev
 
-Then, in another terminal, run :
+Then, _in another terminal,_ run :
 
     $ chimp --chai --ddp=http://localhost:3000
 
@@ -38,6 +53,10 @@ Then, in another terminal, run :
 
 Assertions are done using [Chai](http://chaijs.com/).
 Basically we're throwing errors when something goes awry or against expectations.
+
+``` js
+expect(user.profile.wallet.available).to.equal(expectedVotesAvailable);
+```
 
 
 ## Transformers
@@ -70,5 +89,5 @@ See `setup-steps.js` for an example.
 ### Naming a user 'I'
 
 **Don't.**
-The features will not like it, as they will try to fetch the logged in user defined from previous steps.
+The features will first try to fetch the logged in user defined from previous steps.
